requirements.txt
flask
Flask-SQLAlchemy==2.4.1
psycopg2-binary